% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{association}
\alias{association}
\title{Hodder and Okell's index of association}
\usage{
association(x, y)
}
\arguments{
\item{x}{Coordinates specified as a two-column data frame, matrix, or list
of equal-length vectors}
}
\value{
Hodder and Okell's A index of association between \code{x} and \code{y}. For the
interpretation of this value see Orton (1980, p. 154):
"When the distribution of types A and B are randomly mingled, with no
association or dissociation, A will have a value of about 1: closely
packed but separate distributions have low values of A, and if As and Bs
tend to occur together, then A will be greater than 1."
}
\description{
Calculates Hodder and Okell's (1978) 'A' index of association between two
sets of points, which is the product of the ratios of the average distance
between points within each set to the average distance the sets.
}
\examples{
ceramics <- AZ_A1020_BLM[AZ_A1020_BLM$type == "ceramics",]
stone_tools <- AZ_A1020_BLM[AZ_A1020_BLM$type == "stone tool",]
association(ceramics, stone_tools)
}
\references{
\itemize{
\item Hodder, I. and E. Okell. 1978. A new method for assessing the association
between dismbutions of points in archaeology. In I. Hodder (ed),
\emph{Simulation Studies in Archaeology}, p. 97-107. Cambridge: Cambridge
University Press.
\item Orton, C. 1980. \emph{Mathematics in archaeology}. London: Collins.
}
}
