% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AZ_A1020_BLM}
\alias{AZ_A1020_BLM}
\title{AZ_A1020_BLM Test Data}
\format{
\subsection{\code{AZ_A1020_BLM}}{

A data frame with 358 rows and 3 columns:
\describe{
\item{x}{Easting of location in EPSG ...}
\item{y}{Northing of location in EPSG ...}
\item{type}{Type of Evidence at location}
...
}
}
}
\source{
\if{html}{\out{
<???>
}}
}
\usage{
AZ_A1020_BLM
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
