% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coords.R
\name{coord_matrix}
\alias{coord_matrix}
\title{Coordinates as matrix}
\usage{
coord_matrix(x)
}
\arguments{
\item{x}{Coordinates specified as a two-column data frame, matrix, or list
of equal-length vectors}
}
\description{
Coordinates as matrix
}
\keyword{internal}
