% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\name{distance_matrix}
\alias{distance_matrix}
\title{Distance Matrix}
\usage{
distance_matrix(location_data, x = 1, y = 2)
}
\arguments{
\item{location_data}{data.frame with three columns:
\describe{
\item{x}{Easting / x-coordinate of location}
\item{y}{Northing / y-coordinate of location}
\item{type}{Type of Evidence at location}
}}

\item{x}{Index of column containing the x-coordinate}

\item{y}{Index of column containing the y-coordinate}
}
\description{
The distance_matrix function finds the difference from each point to
every other point, and returns a rectangular matrix (which means each
distance is in the matrix twice, once from point A to point B,
and again from point B to point A)
}
