% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_counts.R
\name{local_count}
\alias{local_count}
\title{Count points within a radius}
\usage{
local_count(x, y, radius)
}
\arguments{
\item{x, }{y Equal-length numeric vectors of coordinates}

\item{radius}{Size of the neighbourhood}
}
\value{
Numeric vector the same length as \code{x} and \code{y} with the number of
points within the specified radius of each point (counting the point itself,
so always >=1).
}
\description{
For a vector of points specified by x and y coordinates, calculates the
number of points within the given radius.
}
\examples{
local_count(AZ_A1020_BLM$x, AZ_A1020_BLM$y, radius = 2)
}
