% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_counts.R
\name{local_counts}
\alias{local_counts}
\title{Title}
\usage{
local_counts(location_data, x = 1, y = 2, radius)
}
\arguments{
\item{location_data}{expects a data.frame with three columns:
\describe{
\item{x}{Easting / x-coordinate of location}
\item{y}{Northing / y-coordinate of location}
\item{type}{categorical variable for type of evidence at location}
}}

\item{x}{Index of column containing the x-coordinate}

\item{y}{Index of column containing the y-coordinate}

\item{radius}{\emph{numeric}. the radius of what constitues the
"neighborhood" of each point}
}
\value{
a data.frame that includes the point location and
type of each point, as well as counts (by type) of points within the
specified radius of each point
}
\description{
Local_counts is a function to find the counts of points of different
types within the local neighborhood of each point -- points are
included in the counts of their own neighborhood.
The output is a dataframe that includes the point location and
type of each point, as well as counts (by type) of points within the
specified radius of each point.
}
\examples{
\dontrun{
  locations <- LDen::AZ_A1020_BLM
  local_counts <- local_counts(locations, radius = 2)
}

}
