% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_density.R
\name{global_density}
\alias{global_density}
\title{Global point density}
\usage{
global_density(x, area = NULL)
}
\arguments{
\item{x}{Coordinates specified as a two-column data frame, matrix, or list
of equal-length vectors}

\item{area}{Total area of the window. If \code{NULL}, inferred as the area of the
bounding box of the coordinates specified by \code{x} and \code{y}.}
}
\value{
Number of points divided by area
}
\description{
Calculates the 'global' density of a vector of point coordinates in relation
to the total size of the window or study region (e.g. the site or trench).
}
\examples{
global_density(AZ_A1020_BLM, area = 2409)
global_density(AZ_A1020_BLM) # area estimated from bounding box
}
