% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{lda}
\alias{lda}
\title{LDA: Local Density ...?}
\usage{
lda(location_data, x = 1, y = 2, radius, site_area)
}
\arguments{
\item{location_data}{expects a data.frame with three columns:
\describe{
\item{x}{Easting / x-coordinate of location}
\item{y}{Northing / y-coordinate of location}
\item{type}{categorical variable for type of evidence at location}
}}

\item{x}{Index of column containing the x-coordinate}

\item{y}{Index of column containing the y-coordinate}

\item{radius}{\emph{numeric}. the radius of what constitues the
"neighborhood" of each point}

\item{site_area}{\emph{numeric}. total area of the site}
}
\value{
A data.frame containing the local density coefficients.
}
\description{
\code{lda()} calculates the local density coefficient within and between
all the types in the dataset. The local density coefficient from
Type A to Type B is the mean density of points of Type B within
the specified radius of points of Type A, divided by the global
density of points of Type B (i.e., the number of points of
Type B divided by the total area)
}
