% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{lda}
\alias{lda}
\title{LDA: Local Density ...?}
\usage{
lda(location_data, radius, site_area)
}
\arguments{
\item{location_data}{data.frame with three columns:
\describe{
\item{x}{Easting / x-coordinate of location}
\item{y}{Northing / y-coordinate of location}
\item{type}{Type of Evidence at location}
}}

\item{radius}{\emph{numeric}. The radius of what}

\item{site_area}{}
}
\description{
\code{lda()} calculates the local density coefficient within and between
all the types in the dataset. The local density coefficient from
Type A to Type B is the mean density of points of Type B within
the specified radius of points of Type A, divided by the global
density of points of Type B (i.e., the number of points of
Type B divided by the total area)
}
