% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_count.R
\name{local_count}
\alias{local_count}
\alias{local_density}
\title{Local point counts and density}
\usage{
local_count(x, y = NULL, radius)

local_density(x, y = NULL, radius)
}
\arguments{
\item{x}{Coordinates specified as a two-column data frame, matrix, or list
of equal-length vectors}

\item{y}{For the two-sample case, coordinates of points to count, in the same
format as \code{x}. Set \code{NULL} (the default) for the one-sample case.}

\item{radius}{Size of the neighbourhood}
}
\value{
Numeric vector the same length as \code{x} with the number or density of
points within the specified radius of each point. In the one sample case,
the count includes the origin point (and thus is always >= 1) but the
density does not.
}
\description{
For a vector of points specified by x and y coordinates, \code{local_count()}
calculates the number of points within the given radius. \code{local_density()}
returns the density of points, i.e. the count divided by the area of a circle
with the given radius.
}
\details{
\code{local_count2()} and \code{local_density2()} are two-sample versions of the above,
calculating the count of density of points in y within the specified radius
of points in x.
}
\examples{
local_count(AZ_A1020_BLM, radius = 2)
local_density(AZ_A1020_BLM, radius = 2)
}
