% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_count.R
\name{local_count}
\alias{local_count}
\alias{local_density}
\alias{local_count2}
\alias{local_density2}
\title{Local point counts and density}
\usage{
local_count(x, y, radius)

local_density(x, y, radius)

local_count2(x1, y1, x2, y2, radius)

local_density2(x1, y1, x2, y2, radius)
}
\arguments{
\item{x, y}{Equal-length numeric vectors of coordinates}

\item{radius}{Size of the neighbourhood}

\item{x1, x2, y1, y2}{Equal-length numeric vectors of coordinates (for two-sample versions)}
}
\value{
Numeric vector the same length as \code{x} with the number (counting the
point itself, so always >=1) or density points within the specified radius of
each point.
}
\description{
For a vector of points specified by x and y coordinates, \code{local_count()}
calculates the number of points within the given radius. \code{local_density()}
returns the density of points, i.e. the count divided by the area of a circle
with the given radius.
}
\details{
\code{local_count2()} and \code{local_density2()} are two-sample versions of the above,
calculating the count of density of points in y within the specified radius
of points in x.
}
\examples{
local_count(AZ_A1020_BLM$x, AZ_A1020_BLM$y, radius = 2)
local_density(AZ_A1020_BLM$x, AZ_A1020_BLM$y, radius = 2)
}
