% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_counts.R
\name{local_counts}
\alias{local_counts}
\title{Title}
\usage{
local_counts(location_data, radius)
}
\arguments{
\item{location_data}{data.frame with three columns:
\describe{
\item{x}{Easting / x-coordinate of location}
\item{y}{Northing / y-coordinate of location}
\item{type}{Type of Evidence at location}
}}

\item{radius}{\emph{numeric}. The radius of what}
}
\description{
Local_counts is a function to find the counts of points of different
types within the local neighborhood of each point -- points are
included in the counts of their own neighborhood.
The output is a dataframe that includes the point location and
type of each point, as well as counts (by type) of points within the
specified radius of each point.
}
\examples{

locations <- LDen::AZ_A1020_BLM
local_counts <- local_counts(locations, radius = 2)

}
