% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_density.R
\name{local_density}
\alias{local_density}
\title{Calculate Local Density}
\usage{
local_density(location_data, radius)
}
\arguments{
\item{location_data}{data.frame with three columns:
\describe{
\item{x}{Easting / x-coordinate of location}
\item{y}{Northing / y-coordinate of location}
\item{type}{Type of Evidence at location}
}}

\item{radius}{}
}
\description{
\code{local_density()} is a function that calls the local_count function,
then calculates the area of the neighborhood defined by the specified
radius. The local_density is the count divided by the area of the
neighborhood. The function outputs a dataframe that includes the point
location and type of each point, as well as the density of points of
each type within the specified radius of each point.
}
\seealso{
\code{\link[=local_count]{local_count()}}
}
